refactor(tools): split misc-handlers.ts into per-tool interaction files
Eliminates the 935-line dumping ground that hosted handlers for 11
unrelated tools. Each tool now owns its interaction code:

  src/tools/fill/fill-interaction.ts
  src/tools/eyedropper/eyedropper-interaction.ts
  src/tools/dodge/dodge-interaction.ts
  src/tools/smudge/smudge-interaction.ts
  src/tools/stamp/stamp-interaction.ts
  src/tools/text/text-interaction.ts        (commitTextEditing too)
  src/tools/crop/crop-interaction.ts        (new dir)
  src/tools/path/path-interaction.ts
  src/tools/shape/shape-interaction.ts
  src/tools/gradient/gradient-interaction.ts

Splits the unioned handleShapeGradientDown into separate handleShapeDown /
handleGradientDown — the dispatch was already a string-tag switch on tool
kind, and registry-style dispatch wants one handler per tool.

Extracts the dab-interpolation helper used by dodge/smudge/stamp into
src/tools/common/dab-interpolation.ts with its own unit tests
(boundary, diagonal, zero-length).

tool-router.ts now imports from per-tool locations. App.tsx,
useKeyboardShortcuts, and TextActionButtons import commitTextEditing
from the text tool directly.
